+++ frontend/src/assets/style.css
@@ -130,7 +130,7 @@ div.mx-auto.v-card.v-sheet.theme--light.rounded-0 {
     margin: auto;
     font-family: 'Hubballi';
     font-weight: 700;
-    font-size: 1.2em;
+    font-size: 1.8em;
 }
 
 .pict_profile {
@@ -147,7 +147,7 @@ div.mx-auto.v-card.v-sheet.theme--light.rounded-0 {
     margin: auto;
     font-family: 'Hubballi';
     font-weight: 700;
-    font-size: 1.1em;
+    font-size: 1.8em;
 }
 
 .app_bar_change_password_profile {
